refactor InvoiceLineService in interface
